+++ src/datagrid/datagrid.utils.js
@@ -156,11 +156,15 @@ export default {
     let parsedHiddenCols;
     let parsedColOrder;
     let isFiltering = false;
-    if (isFilteringData && !grid.disableRememberIsFiltering) {
-      try { isFiltering = JSON.parse(isFilteringData); } catch (e) {
-        // eslint-disable-next-line no-console
-        console.error('Datagrid: error parsing isFilteringData from localStorage', e);
+    if (isFilteringData) {
+      if (!grid.disableRememberIsFiltering) {
+        try { isFiltering = JSON.parse(isFilteringData); } catch (e) {
+          // eslint-disable-next-line no-console
+          console.error('Datagrid: error parsing isFilteringData from localStorage', e);
+        }
       }
+    } else if (grid.defaultShowFilteringRow) {
+      isFiltering = true;
     }
     if (hiddenColumns) {
       try { parsedHiddenCols = JSON.parse(hiddenColumns); } catch (e) {