+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## <next>
 * New `isHidden` property for column, which indicates the default visibility.
 * Add a possibility to use custom config storage instead of the local storage.
+* Add `defaultShowFilteringRow` to `grid` props, allows filtering row to be shown as default
 
 ## 3.0.1
 * Add a checked column to the end of the selected columns list in column settings